code/hello-world/pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer-cli>=0.16.0",
]

[project.scripts]
hw = "hello_world.main:app"